pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evolving-trader"
version = "0.1.0"
description = "A self-evolving multi-agent trading system for A-share markets"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "litellm>=1.50.0",
    "tushare>=1.4.0",
    "pandas>=2.0.0",
    "ta-lib>=0.4.28",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pdfplumber>=0.11.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "apscheduler>=3.10.0",
    "sse-starlette>=2.0.0",
    "PyMySQL>=1.1.0",
    "SQLAlchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
evolving-trader = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
